// frontend/src/api/models/VideoAspectRatio.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Klippers API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const VideoAspectRatio = {
    _169: '16:9',
    _916: '9:16',
    _11: '1:1',
    _45: '4:5',
    _23: '2:3'
} as const;
export type VideoAspectRatio = typeof VideoAspectRatio[keyof typeof VideoAspectRatio];


export function instanceOfVideoAspectRatio(value: any): boolean {
    for (const key in VideoAspectRatio) {
        if (Object.prototype.hasOwnProperty.call(VideoAspectRatio, key)) {
            if (VideoAspectRatio[key as keyof typeof VideoAspectRatio] === value) {
                return true;
            }
        }
    }
    return false;
}

export function VideoAspectRatioFromJSON(json: any): VideoAspectRatio {
    return VideoAspectRatioFromJSONTyped(json, false);
}

export function VideoAspectRatioFromJSONTyped(json: any, ignoreDiscriminator: boolean): VideoAspectRatio {
    return json as VideoAspectRatio;
}

export function VideoAspectRatioToJSON(value?: VideoAspectRatio | null): any {
    return value as any;
}

export function VideoAspectRatioToJSONTyped(value: any, ignoreDiscriminator: boolean): VideoAspectRatio {
    return value as VideoAspectRatio;
}